@@ -15,7 +15,7 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.15.0 <3.0.0"
   flutter: ">=1.10.0"
 
 dependencies:
@@ -23,16 +23,16 @@ dependencies:
     sdk: flutter
   sip_ua:
     path: ../
-  shared_preferences: ^2.3.2
-  permission_handler: ^11.3.1
-  flutter_webrtc: ^0.11.6
-  provider: ^6.1.2
+  shared_preferences: ^2.2.0
+  permission_handler: ^11.1.0
+  flutter_webrtc: ^0.12.4
+  provider: 6.1.2
   logger: ^2.4.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  lints: ^3.0.0
+  lints: ^4.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
