@@ -179,7 +179,7 @@ class _MyRegisterWidget extends State<RegisterWidget>
       borderRadius: BorderRadius.circular(5),
     );
     Color? textLabelColor =
-        Theme.of(context).textTheme.bodyMedium?.color?.withOpacity(0.5);
+        Theme.of(context).textTheme.bodyMedium?.color?.withValues(alpha: 0.5);
     return Scaffold(
       appBar: AppBar(
         title: Text("SIP Account"),
@@ -210,7 +210,7 @@ class _MyRegisterWidget extends State<RegisterWidget>
         children: <Widget>[
           Center(
             child: Text(
-              'Register Status: ${EnumHelper.getName(_registerState.state)}',
+              'Register Status: ${_registerState.state?.name ?? ''}',
               style: TextStyle(fontSize: 18, color: textColor),
             ),
           ),
