@@ -1,16 +1,17 @@
-import 'package:sip_ua/src/sip_message.dart';
 import '../constants.dart';
+import '../enums.dart';
 import '../event_manager/event_manager.dart';
 import '../event_manager/internal_events.dart';
 import '../exceptions.dart' as Exceptions;
-import '../rtc_session.dart' as rtc;
+import '../rtc_session.dart';
+import '../sip_message.dart';
 import '../utils.dart' as utils;
 
 class Info extends EventManager {
   Info(this._session);
 
-  final rtc.RTCSession _session;
-  String? _direction;
+  final RTCSession _session;
+  Direction? _direction;
   String? _contentType;
   String? _body;
   IncomingRequest? _request;
@@ -19,18 +20,18 @@ class Info extends EventManager {
 
   String? get body => _body;
 
-  String? get direction => _direction;
+  Direction? get direction => _direction;
 
   void send(String contentType, String body, Map<String, dynamic> options) {
-    _direction = 'outgoing';
+    _direction = Direction.outgoing;
 
     if (contentType == null) {
       throw Exceptions.TypeError('Not enough arguments');
     }
 
     // Check RTCSession Status.
-    if (_session.status != rtc.C.STATUS_CONFIRMED &&
-        _session.status != rtc.C.STATUS_WAITING_FOR_ACK) {
+    if (_session.state != RtcSessionState.confirmed &&
+        _session.state != RtcSessionState.waitingForAck) {
       throw Exceptions.InvalidStateError(_session.status);
     }
 
@@ -41,14 +42,16 @@ class Info extends EventManager {
 
     extraHeaders.add('Content-Type: $contentType');
 
-    _session.newInfo('local', this, _request);
+    _session.newInfo(Originator.local, this, _request);
 
     EventManager handlers = EventManager();
     handlers.on(EventOnSuccessResponse(), (EventOnSuccessResponse event) {
-      emit(EventSucceeded(originator: 'remote', response: event.response));
+      emit(EventSucceeded(
+          originator: Originator.remote, response: event.response));
     });
     handlers.on(EventOnErrorResponse(), (EventOnErrorResponse event) {
-      emit(EventCallFailed(originator: 'remote', response: event.response));
+      emit(EventCallFailed(
+          originator: Originator.remote, response: event.response));
     });
     handlers.on(EventOnTransportError(), (EventOnTransportError event) {
       _session.onTransportError();
@@ -68,7 +71,7 @@ class Info extends EventManager {
   }
 
   void init_incoming(IncomingRequest request) {
-    _direction = 'incoming';
+    _direction = Direction.incoming;
     _request = request;
 
     request.reply(200);
@@ -76,6 +79,6 @@ class Info extends EventManager {
     _contentType = request.getHeader('content-type');
     _body = request.body;
 
-    _session.newInfo('remote', this, request);
+    _session.newInfo(Originator.remote, this, request);
   }
 }
