@@ -34,7 +34,7 @@ class _MyCallScreenWidget extends State<CallScreenWidget>
   bool _speakerOn = false;
   bool _hold = false;
   bool _mirror = true;
-  String? _holdOriginator;
+  Originator? _holdOriginator;
   bool _callConfirmed = false;
   CallStateEnum _state = CallStateEnum.NONE;
 
@@ -47,7 +47,7 @@ class _MyCallScreenWidget extends State<CallScreenWidget>
 
   String? get remoteIdentity => call!.remote_identity;
 
-  String get direction => call!.direction;
+  Direction? get direction => call!.direction;
 
   Call? get call => widget._call;
 
@@ -177,7 +177,7 @@ class _MyCallScreenWidget extends State<CallScreenWidget>
 
   void _handleStreams(CallState event) async {
     MediaStream? stream = event.stream;
-    if (event.originator == 'local') {
+    if (event.originator == Originator.local) {
       if (_localRenderer != null) {
         _localRenderer!.srcObject = stream;
       }
@@ -189,7 +189,7 @@ class _MyCallScreenWidget extends State<CallScreenWidget>
       }
       _localStream = stream;
     }
-    if (event.originator == 'remote') {
+    if (event.originator == Originator.remote) {
       if (_remoteRenderer != null) {
         _remoteRenderer!.srcObject = stream;
       }
@@ -423,7 +423,7 @@ class _MyCallScreenWidget extends State<CallScreenWidget>
     switch (_state) {
       case CallStateEnum.NONE:
       case CallStateEnum.CONNECTING:
-        if (direction == 'INCOMING') {
+        if (direction == Direction.incoming) {
           basicActions.add(ActionButton(
             title: "Accept",
             fillColor: Colors.green,
@@ -608,7 +608,7 @@ class _MyCallScreenWidget extends State<CallScreenWidget>
                       child: Text(
                         (voiceOnly ? 'VOICE CALL' : 'VIDEO CALL') +
                             (_hold
-                                ? ' PAUSED BY ${_holdOriginator!.toUpperCase()}'
+                                ? ' PAUSED BY ${_holdOriginator!.name}'
                                 : ''),
                         style: TextStyle(fontSize: 24, color: textColor),
                       ),
@@ -655,7 +655,7 @@ class _MyCallScreenWidget extends State<CallScreenWidget>
     return Scaffold(
       appBar: AppBar(
         automaticallyImplyLeading: false,
-        title: Text('[$direction] ${EnumHelper.getName(_state)}'),
+        title: Text('[$direction] ${_state.name}'),
       ),
       body: _buildContent(),
       floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
