@@ -1,7 +1,5 @@
-export 'src/enum_helper.dart';
+export 'src/enums.dart';
 export 'src/sip_message.dart';
 export 'src/sip_ua_helper.dart';
 export 'src/transport_type.dart';
 export 'src/uri.dart';
-
-
