@@ -1,9 +1,10 @@
+import '../enums.dart';
 import '../message.dart';
 import 'events.dart';
 
 class EventNewMessage extends EventType {
   EventNewMessage({this.message, this.originator, this.request});
   dynamic request;
-  String? originator;
+  Originator? originator;
   Message? message;
 }
