@@ -1,4 +1,4 @@
-import 'package:sip_ua/src/sip_message.dart';
+import '../sip_message.dart';
 import 'events.dart';
 
 class EventTerminated extends EventType {
