@@ -1,9 +1,10 @@
+import '../enums.dart';
 import '../options.dart';
 import 'events.dart';
 
 class EventNewOptions extends EventType {
   EventNewOptions({this.message, this.originator, this.request});
   dynamic request;
-  String? originator;
+  Originator? originator;
   Options? message;
 }
