@@ -1,3 +1,4 @@
+import 'package:sip_ua/src/enums.dart';
 import 'package:sip_ua/src/name_addr_header.dart';
 import 'constants.dart' as DartSIP_C;
 import 'constants.dart';
@@ -17,14 +18,14 @@ class Options extends EventManager with Applicant {
   final UA _ua;
   dynamic _request;
   bool _closed = false;
-  String? _direction;
+  Direction? _direction;
   NameAddrHeader? _local_identity;
   NameAddrHeader? _remote_identity;
   // Whether an incoming Options has been replied.
   bool _is_replied = false;
   // Custom Options empty object for high level use.
   final Map<String, dynamic> _data = <String, dynamic>{};
-  String? get direction => _direction;
+  Direction? get direction => _direction;
 
   NameAddrHeader? get local_identity => _local_identity;
 
@@ -75,7 +76,7 @@ class Options extends EventManager with Applicant {
 
     RequestSender request_sender = RequestSender(_ua, _request, handlers);
 
-    _newOptions('local', _request);
+    _newOptions(Originator.local, _request);
 
     request_sender.send();
   }
@@ -83,7 +84,7 @@ class Options extends EventManager with Applicant {
   void init_incoming(IncomingRequest request) {
     _request = request;
 
-    _newOptions('remote', request);
+    _newOptions(Originator.remote, request);
 
     // Reply with a 200 OK if the user didn't reply.
     if (!_is_replied) {
@@ -102,7 +103,7 @@ class Options extends EventManager with Applicant {
     List<dynamic> extraHeaders = Utils.cloneArray(options['extraHeaders']);
     String? body = options['body'];
 
-    if (_direction != 'incoming') {
+    if (_direction != Direction.incoming) {
       throw Exceptions.NotSupportedError(
           '"accept" not supported for outgoing Options');
     }
@@ -125,7 +126,7 @@ class Options extends EventManager with Applicant {
     List<dynamic> extraHeaders = Utils.cloneArray(options['extraHeaders']);
     String? body = options['body'];
 
-    if (_direction != 'incoming') {
+    if (_direction != Direction.incoming) {
       throw Exceptions.NotSupportedError(
           '"reject" not supported for outgoing Options');
     }
@@ -149,10 +150,11 @@ class Options extends EventManager with Applicant {
     if (RegExp(r'^1[0-9]{2}$').hasMatch(response!.status_code)) {
       // Ignore provisional responses.
     } else if (RegExp(r'^2[0-9]{2}$').hasMatch(response.status_code)) {
-      _succeeded('remote', response);
+      _succeeded(Originator.remote, response);
     } else {
       String cause = Utils.sipErrorCause(response.status_code);
-      _failed('remote', response.status_code, cause, response.reason_phrase);
+      _failed(Originator.remote, response.status_code, cause,
+          response.reason_phrase);
     }
   }
 
@@ -160,15 +162,15 @@ class Options extends EventManager with Applicant {
     if (_closed != null) {
       return;
     }
-    _failed(
-        'system', 408, DartSIP_C.CausesType.REQUEST_TIMEOUT, 'Request Timeout');
+    _failed(Originator.system, 408, DartSIP_C.CausesType.REQUEST_TIMEOUT,
+        'Request Timeout');
   }
 
   void _onTransportError() {
     if (_closed != null) {
       return;
     }
-    _failed('system', 500, DartSIP_C.CausesType.CONNECTION_ERROR,
+    _failed(Originator.system, 500, DartSIP_C.CausesType.CONNECTION_ERROR,
         'Transport Error');
   }
 
@@ -182,13 +184,13 @@ class Options extends EventManager with Applicant {
    * Internal Callbacks
    */
 
-  void _newOptions(String originator, dynamic request) {
-    if (originator == 'remote') {
-      _direction = 'incoming';
+  void _newOptions(Originator originator, dynamic request) {
+    if (originator == Originator.remote) {
+      _direction = Direction.incoming;
       _local_identity = request.to;
       _remote_identity = request.from;
-    } else if (originator == 'local') {
-      _direction = 'outgoing';
+    } else if (originator == Originator.local) {
+      _direction = Direction.outgoing;
       _local_identity = request.from;
       _remote_identity = request.to;
     }
@@ -196,7 +198,7 @@ class Options extends EventManager with Applicant {
     _ua.newOptions(this, originator, request);
   }
 
-  void _failed(String originator, int? status_code, String cause,
+  void _failed(Originator originator, int? status_code, String cause,
       String? reason_phrase) {
     logger.d('OPTIONS failed');
     close();
@@ -209,7 +211,7 @@ class Options extends EventManager with Applicant {
             reason_phrase: reason_phrase)));
   }
 
-  void _succeeded(String originator, IncomingResponse? response) {
+  void _succeeded(Originator originator, IncomingResponse? response) {
     logger.d('OPTIONS succeeded');
 
     close();
