@@ -1,12 +1,9 @@
 import 'dart:async';
-import 'dart:convert';
 
-import 'package:crypto/crypto.dart';
 import 'package:flutter_webrtc/flutter_webrtc.dart';
 import 'package:sdp_transform/sdp_transform.dart' as sdp_transform;
-import 'package:sdp_transform/sdp_transform.dart';
 
-import 'package:sip_ua/sip_ua.dart';
+import '../sip_ua.dart';
 import 'constants.dart' as DartSIP_C;
 import 'constants.dart';
 import 'dialog.dart';
@@ -22,25 +19,22 @@ import 'rtc_session/info.dart' as RTCSession_Info;
 import 'rtc_session/info.dart';
 import 'rtc_session/refer_notifier.dart';
 import 'rtc_session/refer_subscriber.dart';
-import 'sip_message.dart';
 import 'timers.dart';
 import 'transactions/transaction_base.dart';
 import 'ua.dart';
-import 'uri.dart';
 import 'utils.dart' as utils;
 
-class C {
-  // RTCSession states.
-  static const int STATUS_NULL = 0;
-  static const int STATUS_INVITE_SENT = 1;
-  static const int STATUS_1XX_RECEIVED = 2;
-  static const int STATUS_INVITE_RECEIVED = 3;
-  static const int STATUS_WAITING_FOR_ANSWER = 4;
-  static const int STATUS_ANSWERED = 5;
-  static const int STATUS_WAITING_FOR_ACK = 6;
-  static const int STATUS_CANCELED = 7;
-  static const int STATUS_TERMINATED = 8;
-  static const int STATUS_CONFIRMED = 9;
+enum RtcSessionState {
+  none, // STATUS_NULL
+  inviteSent, // STATUS_INVITE_SENT
+  provisionalResponse, // STATUS_1XX_RECEIVED
+  inviteReceived, // STATUS_INVITE_RECEIVED
+  waitingForAnswer, // STATUS_WAITING_FOR_ANSWER
+  answered, // STATUS_ANSWERED
+  waitingForAck, // STATUS_WAITING_FOR_ACK
+  canceled, // STATUS_CANCELED
+  terminated, // STATUS_TERMINATED
+  confirmed, // STATUS_CONFIRMED
 }
 
 /**
@@ -86,7 +80,7 @@ class RTCSession extends EventManager implements Owner {
   final UA _ua;
 
   String? _id;
-  int _status = C.STATUS_NULL;
+  RtcSessionState _state = RtcSessionState.none;
   Dialog? _dialog;
   final Map<String?, Dialog> _earlyDialogs = <String?, Dialog>{};
   String? _contact;
@@ -126,7 +120,7 @@ class RTCSession extends EventManager implements Owner {
   final SIPTimers _timers = SIPTimers();
 
   // Session info.
-  String? _direction;
+  Direction? _direction;
   NameAddrHeader? _local_identity;
   NameAddrHeader? _remote_identity;
   DateTime? _start_time;
@@ -169,14 +163,14 @@ class RTCSession extends EventManager implements Owner {
   RTCPeerConnection? get connection => _connection;
 
   @override
-  int get TerminatedCode => C.STATUS_TERMINATED;
+  int get TerminatedCode => RtcSessionState.terminated.index;
 
   RTCDTMFSender get dtmfSender =>
       _connection!.createDtmfSender(_localMediaStream!.getAudioTracks()[0]);
 
   String? get contact => _contact;
 
-  String? get direction => _direction;
+  Direction? get direction => _direction;
 
   NameAddrHeader? get local_identity => _local_identity;
 
@@ -190,15 +184,17 @@ class RTCSession extends EventManager implements Owner {
   UA get ua => _ua;
 
   @override
-  int get status => _status;
+  int get status => _state.index;
+
+  RtcSessionState get state => _state;
 
   bool isInProgress() {
-    switch (_status) {
-      case C.STATUS_NULL:
-      case C.STATUS_INVITE_SENT:
-      case C.STATUS_1XX_RECEIVED:
-      case C.STATUS_INVITE_RECEIVED:
-      case C.STATUS_WAITING_FOR_ANSWER:
+    switch (_state) {
+      case RtcSessionState.none:
+      case RtcSessionState.inviteSent:
+      case RtcSessionState.provisionalResponse:
+      case RtcSessionState.inviteReceived:
+      case RtcSessionState.waitingForAnswer:
         return true;
       default:
         return false;
@@ -206,10 +202,10 @@ class RTCSession extends EventManager implements Owner {
   }
 
   bool isEstablished() {
-    switch (_status) {
-      case C.STATUS_ANSWERED:
-      case C.STATUS_WAITING_FOR_ACK:
-      case C.STATUS_CONFIRMED:
+    switch (_state) {
+      case RtcSessionState.answered:
+      case RtcSessionState.waitingForAck:
+      case RtcSessionState.confirmed:
         return true;
       default:
         return false;
@@ -217,9 +213,9 @@ class RTCSession extends EventManager implements Owner {
   }
 
   bool isEnded() {
-    switch (_status) {
-      case C.STATUS_CANCELED:
-      case C.STATUS_TERMINATED:
+    switch (_state) {
+      case RtcSessionState.canceled:
+      case RtcSessionState.terminated:
         return true;
       default:
         return false;
@@ -264,8 +260,8 @@ class RTCSession extends EventManager implements Owner {
     }
 
     // Check Session Status.
-    if (_status != C.STATUS_NULL) {
-      throw Exceptions.InvalidStateError(_status);
+    if (_state != RtcSessionState.none) {
+      throw Exceptions.InvalidStateError(_state.name);
     }
 
     // Check WebRTC support.
@@ -341,7 +337,7 @@ class RTCSession extends EventManager implements Owner {
     await _createRTCConnection(pcConfig, rtcConstraints);
 
     // Set internal properties.
-    _direction = 'outgoing';
+    _direction = Direction.outgoing;
     _local_identity = _request.from;
     _remote_identity = _request.to;
 
@@ -350,7 +346,7 @@ class RTCSession extends EventManager implements Owner {
       initCallback(this);
     }
 
-    _newRTCSession('local', _request);
+    _newRTCSession(Originator.local, _request);
     await _sendInitialRequest(
         pcConfig, mediaConstraints, rtcOfferConstraints, mediaStream);
   }
@@ -369,7 +365,7 @@ class RTCSession extends EventManager implements Owner {
     }
 
     // Session parameter initialization.
-    _status = C.STATUS_INVITE_RECEIVED;
+    _state = RtcSessionState.inviteReceived;
     _from_tag = request.from_tag;
     _id = request.call_id! + _from_tag!;
     _request = request;
@@ -405,13 +401,13 @@ class RTCSession extends EventManager implements Owner {
       _late_sdp = true;
     }
 
-    _status = C.STATUS_WAITING_FOR_ANSWER;
+    _state = RtcSessionState.waitingForAnswer;
 
     // Set userNoAnswerTimer.
     _timers.userNoAnswerTimer = setTimeout(() {
       request.reply(408);
-      _failed('local', null, null, null, 408, DartSIP_C.CausesType.NO_ANSWER,
-          'No Answer');
+      _failed(Originator.local, null, null, null, 408,
+          DartSIP_C.CausesType.NO_ANSWER, 'No Answer');
     }, _ua.configuration.no_answer_timeout);
 
     /* Set expiresTimer
@@ -419,16 +415,16 @@ class RTCSession extends EventManager implements Owner {
      */
     if (expires != null) {
       _timers.expiresTimer = setTimeout(() {
-        if (_status == C.STATUS_WAITING_FOR_ANSWER) {
+        if (_state == RtcSessionState.waitingForAnswer) {
           request.reply(487);
-          _failed('system', null, null, null, 487, DartSIP_C.CausesType.EXPIRES,
-              'Timeout');
+          _failed(Originator.system, null, null, null, 487,
+              DartSIP_C.CausesType.EXPIRES, 'Timeout');
         }
       }, expires);
     }
 
     // Set internal properties.
-    _direction = 'incoming';
+    _direction = Direction.incoming;
     _local_identity = request.to;
     _remote_identity = request.from;
 
@@ -438,10 +434,10 @@ class RTCSession extends EventManager implements Owner {
     }
 
     // Fire 'newRTCSession' event.
-    _newRTCSession('remote', request);
+    _newRTCSession(Originator.remote, request);
 
     // The user may have rejected the call in the 'newRTCSession' event.
-    if (_status == C.STATUS_TERMINATED) {
+    if (_state == RtcSessionState.terminated) {
       return;
     }
 
@@ -450,7 +446,7 @@ class RTCSession extends EventManager implements Owner {
 
     // Fire 'progress' event.
     // TODO(cloudwebrtc): Document that 'response' field in 'progress' event is null for incoming calls.
-    _progress('local', null);
+    _progress(Originator.local, null);
   }
 
   /**
@@ -488,14 +484,14 @@ class RTCSession extends EventManager implements Owner {
     data = options['data'] ?? data;
 
     // Check Session Direction and Status.
-    if (_direction != 'incoming') {
+    if (_direction != Direction.incoming) {
       throw Exceptions.NotSupportedError(
           '"answer" not supported for outgoing RTCSession');
     }
 
     // Check Session status.
-    if (_status != C.STATUS_WAITING_FOR_ANSWER) {
-      throw Exceptions.InvalidStateError(_status);
+    if (_state != RtcSessionState.waitingForAnswer) {
+      throw Exceptions.InvalidStateError(_state.name);
     }
 
     // Session Timers.
@@ -509,7 +505,7 @@ class RTCSession extends EventManager implements Owner {
       }
     }
 
-    _status = C.STATUS_ANSWERED;
+    _state = RtcSessionState.answered;
 
     // An error on dialog creation will fire 'failed' event.
     if (!_createDialog(request, 'UAS')) {
@@ -589,7 +585,8 @@ class RTCSession extends EventManager implements Owner {
     // A local MediaStream is given, use it.
     if (mediaStream != null) {
       stream = mediaStream;
-      emit(EventStream(session: this, originator: 'local', stream: stream));
+      emit(EventStream(
+          session: this, originator: Originator.local, stream: stream));
     }
     // Audio and/or video requested, prompt getUserMedia.
     else if (mediaConstraints['audio'] != null ||
@@ -597,14 +594,15 @@ class RTCSession extends EventManager implements Owner {
       _localMediaStreamLocallyGenerated = true;
       try {
         stream = await navigator.mediaDevices.getUserMedia(mediaConstraints);
-        emit(EventStream(session: this, originator: 'local', stream: stream));
+        emit(EventStream(
+            session: this, originator: Originator.local, stream: stream));
       } catch (error) {
-        if (_status == C.STATUS_TERMINATED) {
+        if (_state == RtcSessionState.terminated) {
           throw Exceptions.InvalidStateError('terminated');
         }
         request.reply(480);
         _failed(
-            'local',
+            Originator.local,
             null,
             null,
             null,
@@ -617,7 +615,7 @@ class RTCSession extends EventManager implements Owner {
       }
     }
 
-    if (_status == C.STATUS_TERMINATED) {
+    if (_state == RtcSessionState.terminated) {
       throw Exceptions.InvalidStateError('terminated');
     }
 
@@ -645,15 +643,17 @@ class RTCSession extends EventManager implements Owner {
 
     logger.d('emit "sdp"');
     final String? processedSDP = _sdpOfferToWebRTC(request.body);
-    emit(EventSdp(originator: 'remote', type: 'offer', sdp: processedSDP));
+    emit(EventSdp(
+        originator: Originator.remote, type: SdpType.offer, sdp: processedSDP));
 
-    RTCSessionDescription offer = RTCSessionDescription(processedSDP, 'offer');
+    RTCSessionDescription offer =
+        RTCSessionDescription(processedSDP, SdpType.offer.name);
     try {
       await _connection!.setRemoteDescription(offer);
     } catch (error) {
       request.reply(488);
       _failed(
-          'system',
+          Originator.system,
           null,
           null,
           null,
@@ -668,7 +668,7 @@ class RTCSession extends EventManager implements Owner {
     }
 
     // Create local description.
-    if (_status == C.STATUS_TERMINATED) {
+    if (_state == RtcSessionState.terminated) {
       throw Exceptions.InvalidStateError('terminated');
     }
 
@@ -677,16 +677,18 @@ class RTCSession extends EventManager implements Owner {
     RTCSessionDescription desc;
     try {
       if (!_late_sdp) {
-        desc = await _createLocalDescription('answer', rtcAnswerConstraints);
+        desc =
+            await _createLocalDescription(SdpType.answer, rtcAnswerConstraints);
       } else {
-        desc = await _createLocalDescription('offer', _rtcOfferConstraints);
+        desc =
+            await _createLocalDescription(SdpType.offer, _rtcOfferConstraints);
       }
     } catch (e) {
       request.reply(500);
       throw Exceptions.TypeError('_createLocalDescription() failed');
     }
 
-    if (_status == C.STATUS_TERMINATED) {
+    if (_state == RtcSessionState.terminated) {
       throw Exceptions.InvalidStateError('terminated');
     }
 
@@ -694,16 +696,16 @@ class RTCSession extends EventManager implements Owner {
     try {
       _handleSessionTimersInIncomingRequest(request, extraHeaders);
       request.reply(200, null, extraHeaders, desc.sdp, () {
-        _status = C.STATUS_WAITING_FOR_ACK;
+        _state = RtcSessionState.waitingForAck;
         _setInvite2xxTimer(request, desc.sdp);
         _setACKTimer();
-        _accepted('local');
+        _accepted(Originator.local);
       }, () {
-        _failed('system', null, null, null, 500,
+        _failed(Originator.system, null, null, null, 500,
             DartSIP_C.CausesType.CONNECTION_ERROR, 'Transport Error');
       });
     } catch (error, s) {
-      if (_status == C.STATUS_TERMINATED) {
+      if (_state == RtcSessionState.terminated) {
         return;
       }
       logger.e('Failed to answer(): ${error.toString()}',
@@ -731,15 +733,15 @@ class RTCSession extends EventManager implements Owner {
     String? reason_phrase = options['reason_phrase'] as String?;
 
     // Check Session Status.
-    if (_status == C.STATUS_TERMINATED) {
-      throw Exceptions.InvalidStateError(_status);
+    if (_state == RtcSessionState.terminated) {
+      throw Exceptions.InvalidStateError(_state.name);
     }
 
-    switch (_status) {
+    switch (_state) {
       // - UAC -
-      case C.STATUS_NULL:
-      case C.STATUS_INVITE_SENT:
-      case C.STATUS_1XX_RECEIVED:
+      case RtcSessionState.none:
+      case RtcSessionState.inviteSent:
+      case RtcSessionState.provisionalResponse:
         logger.d('canceling session');
 
         if (status_code != null && (status_code < 200 || status_code >= 700)) {
@@ -750,23 +752,24 @@ class RTCSession extends EventManager implements Owner {
         }
 
         // Check Session Status.
-        if (_status == C.STATUS_NULL || _status == C.STATUS_INVITE_SENT) {
+        if (_state == RtcSessionState.none ||
+            _state == RtcSessionState.inviteSent) {
           _is_canceled = true;
           _cancel_reason = cancel_reason;
-        } else if (_status == C.STATUS_1XX_RECEIVED) {
+        } else if (_state == RtcSessionState.provisionalResponse) {
           _request.cancel(cancel_reason ?? '');
         }
 
-        _status = C.STATUS_CANCELED;
+        _state = RtcSessionState.canceled;
         cancel_reason = cancel_reason ?? 'Canceled by local';
         status_code = status_code ?? 100;
-        _failed('local', null, null, null, status_code,
+        _failed(Originator.local, null, null, null, status_code,
             DartSIP_C.CausesType.CANCELED, cancel_reason);
-        break;
 
+        break;
       // - UAS -
-      case C.STATUS_WAITING_FOR_ANSWER:
-      case C.STATUS_ANSWERED:
+      case RtcSessionState.waitingForAnswer:
+      case RtcSessionState.answered:
         logger.d('rejecting session');
 
         status_code = status_code ?? 480;
@@ -777,12 +780,11 @@ class RTCSession extends EventManager implements Owner {
         }
 
         _request.reply(status_code, reason_phrase, extraHeaders, body);
-        _failed('local', null, null, null, status_code,
+        _failed(Originator.local, null, null, null, status_code,
             DartSIP_C.CausesType.REJECTED, reason_phrase);
         break;
-
-      case C.STATUS_WAITING_FOR_ACK:
-      case C.STATUS_CONFIRMED:
+      case RtcSessionState.waitingForAck:
+      case RtcSessionState.confirmed:
         logger.d('terminating session');
 
         reason_phrase = options['reason_phrase'] as String? ??
@@ -802,8 +804,8 @@ class RTCSession extends EventManager implements Owner {
           * until it has received an ACK for its 2xx response or until the server
           * transaction times out."
           */
-        if (_status == C.STATUS_WAITING_FOR_ACK &&
-            _direction == 'incoming' &&
+        if (_state == RtcSessionState.waitingForAck &&
+            _direction == Direction.incoming &&
             _request.server_transaction.state != TransactionState.TERMINATED) {
           /// Save the dialog for later restoration.
           Dialog dialog = _dialog!;
@@ -811,10 +813,10 @@ class RTCSession extends EventManager implements Owner {
           // Send the BYE as soon as the ACK is received...
           receiveRequest = (IncomingMessage request) {
             if (request.method == SipMethod.ACK) {
-              sendRequest(SipMethod.BYE, <String, dynamic>{
-                'extraHeaders': extraHeaders,
-                'body': body
-              });
+              sendRequest(
+                SipMethod.BYE,
+                <String, dynamic>{'extraHeaders': extraHeaders, 'body': body},
+              );
               dialog.terminate();
             }
           };
@@ -833,12 +835,13 @@ class RTCSession extends EventManager implements Owner {
           });
 
           _ended(
-              'local',
+              Originator.local,
               null,
               ErrorCause(
-                  cause: cause as String?,
-                  status_code: status_code,
-                  reason_phrase: reason_phrase));
+                cause: cause as String?,
+                status_code: status_code,
+                reason_phrase: reason_phrase,
+              ));
 
           // Restore the dialog into 'this' in order to be able to send the in-dialog BYE :-).
           _dialog = dialog;
@@ -851,13 +854,16 @@ class RTCSession extends EventManager implements Owner {
           reason_phrase = reason_phrase ?? 'Terminated by local';
           status_code = status_code ?? 200;
           _ended(
-              'local',
+              Originator.local,
               null,
               ErrorCause(
                   cause: cause as String?,
                   status_code: status_code,
                   reason_phrase: reason_phrase));
         }
+        break;
+      default:
+        break;
     }
   }
 
@@ -880,8 +886,9 @@ class RTCSession extends EventManager implements Owner {
     }
 
     // Check Session Status.
-    if (_status != C.STATUS_CONFIRMED && _status != C.STATUS_WAITING_FOR_ACK) {
-      throw Exceptions.InvalidStateError(_status);
+    if (_state != RtcSessionState.confirmed &&
+        _state != RtcSessionState.waitingForAck) {
+      throw Exceptions.InvalidStateError(_state);
     }
 
     // Convert to string.
@@ -934,7 +941,7 @@ class RTCSession extends EventManager implements Owner {
       if (tone == ',') {
         // queue the delay
         dtmfFuture = dtmfFuture.then((_) async {
-          if (_status == C.STATUS_TERMINATED) {
+          if (_state == RtcSessionState.terminated) {
             return;
           }
           await Future<void>.delayed(Duration(milliseconds: 2000), () {});
@@ -942,7 +949,7 @@ class RTCSession extends EventManager implements Owner {
       } else {
         // queue playing the tone
         dtmfFuture = dtmfFuture.then((_) async {
-          if (_status == C.STATUS_TERMINATED) {
+          if (_state == RtcSessionState.terminated) {
             return;
           }
 
@@ -967,8 +974,9 @@ class RTCSession extends EventManager implements Owner {
     logger.d('sendInfo()');
 
     // Check Session Status.
-    if (_status != C.STATUS_CONFIRMED && _status != C.STATUS_WAITING_FOR_ACK) {
-      throw Exceptions.InvalidStateError(_status);
+    if (_state != RtcSessionState.confirmed &&
+        _state != RtcSessionState.waitingForAck) {
+      throw Exceptions.InvalidStateError(_state);
     }
 
     RTCSession_Info.Info info = RTCSession_Info.Info(this);
@@ -1038,7 +1046,8 @@ class RTCSession extends EventManager implements Owner {
 
     options = options ?? <String, dynamic>{};
 
-    if (_status != C.STATUS_WAITING_FOR_ACK && _status != C.STATUS_CONFIRMED) {
+    if (_state != RtcSessionState.waitingForAck &&
+        _state != RtcSessionState.confirmed) {
       return false;
     }
 
@@ -1051,7 +1060,7 @@ class RTCSession extends EventManager implements Owner {
     }
 
     _localHold = true;
-    _onhold('local');
+    _onhold(Originator.local);
 
     EventManager handlers = EventManager();
 
@@ -1090,7 +1099,8 @@ class RTCSession extends EventManager implements Owner {
 
     options = options ?? <String, dynamic>{};
 
-    if (_status != C.STATUS_WAITING_FOR_ACK && _status != C.STATUS_CONFIRMED) {
+    if (_state != RtcSessionState.waitingForAck &&
+        _state != RtcSessionState.confirmed) {
       return false;
     }
 
@@ -1103,7 +1113,7 @@ class RTCSession extends EventManager implements Owner {
     }
 
     _localHold = false;
-    _onunhold('local');
+    _onunhold(Originator.local);
 
     EventManager handlers = EventManager();
     handlers.on(EventSucceeded(), (EventSucceeded event) {
@@ -1153,7 +1163,8 @@ class RTCSession extends EventManager implements Owner {
     dynamic sdpSemantics =
         options['pcConfig']?['sdpSemantics'] ?? 'unified-plan';
 
-    if (_status != C.STATUS_WAITING_FOR_ACK && _status != C.STATUS_CONFIRMED) {
+    if (_state != RtcSessionState.waitingForAck &&
+        _state != RtcSessionState.confirmed) {
       return false;
     }
 
@@ -1224,7 +1235,8 @@ class RTCSession extends EventManager implements Owner {
 
     dynamic originalTarget = target;
 
-    if (_status != C.STATUS_WAITING_FOR_ACK && _status != C.STATUS_CONFIRMED) {
+    if (_state != RtcSessionState.waitingForAck &&
+        _state != RtcSessionState.confirmed) {
       return null;
     }
 
@@ -1286,22 +1298,22 @@ class RTCSession extends EventManager implements Owner {
       * Terminate the whole session in case the user didn't accept (or yet send the answer)
       * nor reject the request opening the session.
       */
-      if (_status == C.STATUS_WAITING_FOR_ANSWER ||
-          _status == C.STATUS_ANSWERED) {
-        _status = C.STATUS_CANCELED;
+      if (_state == RtcSessionState.waitingForAnswer ||
+          _state == RtcSessionState.answered) {
+        _state = RtcSessionState.canceled;
         _request.reply(487);
-        _failed('remote', null, request, null, 487,
+        _failed(Originator.remote, null, request, null, 487,
             DartSIP_C.CausesType.CANCELED, request.reason_phrase);
       }
     } else {
       // Requests arriving here are in-dialog requests.
       switch (request.method) {
         case SipMethod.ACK:
-          if (_status != C.STATUS_WAITING_FOR_ACK) {
+          if (_state != RtcSessionState.waitingForAck) {
             return;
           }
           // Update signaling status.
-          _status = C.STATUS_CONFIRMED;
+          _state = RtcSessionState.confirmed;
           clearTimeout(_timers.ackTimer);
           clearTimeout(_timers.invite2xxTimer);
 
@@ -1316,10 +1328,12 @@ class RTCSession extends EventManager implements Owner {
 
             logger.d('emit "sdp"');
             emit(EventSdp(
-                originator: 'remote', type: 'answer', sdp: request.body));
+                originator: Originator.remote,
+                type: SdpType.answer,
+                sdp: request.body));
 
             RTCSessionDescription answer =
-                RTCSessionDescription(request.body, 'answer');
+                RTCSessionDescription(request.body, SdpType.answer.name);
             try {
               await _connection!.setRemoteDescription(answer);
             } catch (error) {
@@ -1333,24 +1347,24 @@ class RTCSession extends EventManager implements Owner {
             }
           }
           if (!_is_confirmed) {
-            _confirmed('remote', request);
+            _confirmed(Originator.remote, request);
           }
           break;
         case SipMethod.BYE:
-          if (_status == C.STATUS_CONFIRMED) {
+          if (_state == RtcSessionState.confirmed) {
             request.reply(200);
             _ended(
-                'remote',
+                Originator.remote,
                 request,
                 ErrorCause(
                     cause: DartSIP_C.CausesType.BYE,
                     status_code: 200,
                     reason_phrase: 'BYE Received'));
-          } else if (_status == C.STATUS_INVITE_RECEIVED) {
+          } else if (_state == RtcSessionState.inviteReceived) {
             request.reply(200);
             _request.reply(487, 'BYE Received');
             _ended(
-                'remote',
+                Originator.remote,
                 request,
                 ErrorCause(
                     cause: DartSIP_C.CausesType.BYE,
@@ -1361,7 +1375,7 @@ class RTCSession extends EventManager implements Owner {
           }
           break;
         case SipMethod.INVITE:
-          if (_status == C.STATUS_CONFIRMED) {
+          if (_state == RtcSessionState.confirmed) {
             if (request.hasHeader('replaces')) {
               _receiveReplaces(request);
             } else {
@@ -1372,11 +1386,11 @@ class RTCSession extends EventManager implements Owner {
           }
           break;
         case SipMethod.INFO:
-          if (_status == C.STATUS_1XX_RECEIVED ||
-              _status == C.STATUS_WAITING_FOR_ANSWER ||
-              _status == C.STATUS_ANSWERED ||
-              _status == C.STATUS_WAITING_FOR_ACK ||
-              _status == C.STATUS_CONFIRMED) {
+          if (_state == RtcSessionState.provisionalResponse ||
+              _state == RtcSessionState.waitingForAnswer ||
+              _state == RtcSessionState.answered ||
+              _state == RtcSessionState.waitingForAck ||
+              _state == RtcSessionState.confirmed) {
             String? contentType = request.getHeader('content-type');
             if (contentType != null &&
                 contentType.contains(RegExp(r'^application\/dtmf-relay',
@@ -1392,21 +1406,21 @@ class RTCSession extends EventManager implements Owner {
           }
           break;
         case SipMethod.UPDATE:
-          if (_status == C.STATUS_CONFIRMED) {
+          if (_state == RtcSessionState.confirmed) {
             _receiveUpdate(request);
           } else {
             request.reply(403, 'Wrong Status');
           }
           break;
         case SipMethod.REFER:
-          if (_status == C.STATUS_CONFIRMED) {
+          if (_state == RtcSessionState.confirmed) {
             _receiveRefer(request);
           } else {
             request.reply(403, 'Wrong Status');
           }
           break;
         case SipMethod.NOTIFY:
-          if (_status == C.STATUS_CONFIRMED) {
+          if (_state == RtcSessionState.confirmed) {
             _receiveNotify(request);
           } else {
             request.reply(403, 'Wrong Status');
@@ -1423,7 +1437,7 @@ class RTCSession extends EventManager implements Owner {
    */
   void onTransportError() {
     logger.e('onTransportError()');
-    if (_status != C.STATUS_TERMINATED) {
+    if (_state != RtcSessionState.terminated) {
       terminate(<String, dynamic>{
         'status_code': 500,
         'reason_phrase': DartSIP_C.CausesType.CONNECTION_ERROR,
@@ -1435,7 +1449,7 @@ class RTCSession extends EventManager implements Owner {
   void onRequestTimeout() {
     logger.e('onRequestTimeout()');
 
-    if (_status != C.STATUS_TERMINATED) {
+    if (_state != RtcSessionState.terminated) {
       terminate(<String, dynamic>{
         'status_code': 408,
         'reason_phrase': DartSIP_C.CausesType.REQUEST_TIMEOUT,
@@ -1447,7 +1461,7 @@ class RTCSession extends EventManager implements Owner {
   void onDialogError() {
     logger.e('onDialogError()');
 
-    if (_status != C.STATUS_TERMINATED) {
+    if (_state != RtcSessionState.terminated) {
       terminate(<String, dynamic>{
         'status_code': 500,
         'reason_phrase': DartSIP_C.CausesType.DIALOG_ERROR,
@@ -1457,14 +1471,14 @@ class RTCSession extends EventManager implements Owner {
   }
 
   // Called from DTMF handler.
-  void newDTMF(String originator, DTMF dtmf, dynamic request) {
+  void newDTMF(Originator originator, DTMF dtmf, dynamic request) {
     logger.d('newDTMF()');
 
     emit(EventNewDTMF(originator: originator, dtmf: dtmf, request: request));
   }
 
   // Called from Info handler.
-  void newInfo(String originator, Info info, dynamic request) {
+  void newInfo(Originator originator, Info info, dynamic request) {
     logger.d('newInfo()');
 
     emit(EventNewInfo(originator: originator, info: info, request: request));
@@ -1501,10 +1515,10 @@ class RTCSession extends EventManager implements Owner {
 
   void _close() async {
     logger.d('close()');
-    if (_status == C.STATUS_TERMINATED) {
+    if (_state == RtcSessionState.terminated) {
       return;
     }
-    _status = C.STATUS_TERMINATED;
+    _state = RtcSessionState.terminated;
     // Terminate RTC.
     if (_connection != null) {
       try {
@@ -1565,7 +1579,7 @@ class RTCSession extends EventManager implements Owner {
     int timeout = Timers.T1;
 
     void invite2xxRetransmission() {
-      if (_status != C.STATUS_WAITING_FOR_ACK) {
+      if (_state != RtcSessionState.waitingForAck) {
         return;
       }
       request.reply(200, null, <String>['Contact: $_contact'], body);
@@ -1588,13 +1602,13 @@ class RTCSession extends EventManager implements Owner {
    */
   void _setACKTimer() {
     _timers.ackTimer = setTimeout(() {
-      if (_status == C.STATUS_WAITING_FOR_ACK) {
+      if (_state == RtcSessionState.waitingForAck) {
         logger.d('no ACK received, terminating the session');
 
         clearTimeout(_timers.invite2xxTimer);
         sendRequest(SipMethod.BYE);
         _ended(
-            'remote',
+            Originator.remote,
             null,
             ErrorCause(
                 cause: DartSIP_C.CausesType.NO_ACK,
@@ -1642,14 +1656,16 @@ class RTCSession extends EventManager implements Owner {
         _connection!.onTrack = (RTCTrackEvent event) {
           if (event.streams.isNotEmpty) {
             emit(EventStream(
-                session: this, originator: 'remote', stream: event.streams[0]));
+                session: this,
+                originator: Originator.remote,
+                stream: event.streams[0]));
           }
         };
         break;
       case 'plan-b':
         _connection!.onAddStream = (MediaStream stream) {
-          emit(
-              EventStream(session: this, originator: 'remote', stream: stream));
+          emit(EventStream(
+              session: this, originator: Originator.remote, stream: stream));
         };
         break;
     }
@@ -1660,7 +1676,7 @@ class RTCSession extends EventManager implements Owner {
   }
 
   Future<RTCSessionDescription> _createLocalDescription(
-      String type, Map<String, dynamic>? constraints) async {
+      SdpType type, Map<String, dynamic>? constraints) async {
     logger.d('createLocalDescription()');
     _iceGatheringState ??= RTCIceGatheringState.RTCIceGatheringStateNew;
     Completer<RTCSessionDescription> completer =
@@ -1678,14 +1694,14 @@ class RTCSession extends EventManager implements Owner {
 
     constraints['offerModifiers'] = null;
 
-    if (type != 'offer' && type != 'answer') {
+    if (type != SdpType.offer && type != SdpType.answer) {
       completer.completeError(Exceptions.TypeError(
           'createLocalDescription() | invalid type "$type"'));
     }
 
     _rtcReady = false;
     late RTCSessionDescription desc;
-    if (type == 'offer') {
+    if (type == SdpType.offer) {
       try {
         desc = await _connection!.createOffer(constraints);
       } catch (error) {
@@ -1714,7 +1730,7 @@ class RTCSession extends EventManager implements Owner {
     }
 
     Future<void> ready() async {
-      if (!finished && _status != C.STATUS_TERMINATED) {
+      if (!finished && _state != RtcSessionState.terminated) {
         finished = true;
         _connection!.onIceCandidate = null;
         _connection!.onIceGatheringState = null;
@@ -1722,7 +1738,8 @@ class RTCSession extends EventManager implements Owner {
         _rtcReady = true;
         RTCSessionDescription? desc = await _connection!.getLocalDescription();
         logger.d('emit "sdp"');
-        emit(EventSdp(originator: 'local', type: type, sdp: desc!.sdp));
+        emit(
+            EventSdp(originator: Originator.local, type: type, sdp: desc!.sdp));
         completer.complete(desc);
       }
     }
@@ -1769,7 +1786,7 @@ class RTCSession extends EventManager implements Owner {
       _rtcReady = true;
       RTCSessionDescription? desc = await _connection!.getLocalDescription();
       logger.d('emit "sdp"');
-      emit(EventSdp(originator: 'local', type: type, sdp: desc!.sdp));
+      emit(EventSdp(originator: Originator.local, type: type, sdp: desc!.sdp));
       return desc;
     }
 
@@ -1795,7 +1812,7 @@ class RTCSession extends EventManager implements Owner {
         } catch (error) {
           logger.d('$error');
           _failed(
-              'remote',
+              Originator.remote,
               message,
               null,
               null,
@@ -1828,7 +1845,7 @@ class RTCSession extends EventManager implements Owner {
       } catch (error) {
         logger.d(error.toString());
         _failed(
-            'remote',
+            Originator.remote,
             message,
             null,
             null,
@@ -1855,7 +1872,7 @@ class RTCSession extends EventManager implements Owner {
       }
 
       request.reply(200, null, extraHeaders, sdp, () {
-        _status = C.STATUS_WAITING_FOR_ACK;
+        _state = RtcSessionState.waitingForAck;
         _setInvite2xxTimer(request, sdp);
         _setACKTimer();
       });
@@ -1874,7 +1891,7 @@ class RTCSession extends EventManager implements Owner {
 
       try {
         RTCSessionDescription desc =
-            await _createLocalDescription('offer', _rtcOfferConstraints);
+            await _createLocalDescription(SdpType.offer, _rtcOfferConstraints);
         sendAnswer(desc.sdp);
       } catch (_) {
         request.reply(500);
@@ -1894,7 +1911,7 @@ class RTCSession extends EventManager implements Owner {
       String? reason_phrase = options['reason_phrase'];
       List<dynamic> extraHeaders = utils.cloneArray(options['extraHeaders']);
 
-      if (_status != C.STATUS_CONFIRMED) {
+      if (_state != RtcSessionState.confirmed) {
         return false;
       }
 
@@ -1911,7 +1928,7 @@ class RTCSession extends EventManager implements Owner {
     Future<bool> acceptReInvite(dynamic options) async {
       try {
         // Send answer.
-        if (_status == C.STATUS_TERMINATED) {
+        if (_state == RtcSessionState.terminated) {
           return false;
         }
         sendAnswer(desc.sdp);
@@ -1959,7 +1976,7 @@ class RTCSession extends EventManager implements Owner {
       String reason_phrase = options['reason_phrase'] ?? '';
       List<dynamic> extraHeaders = utils.cloneArray(options['extraHeaders']);
 
-      if (_status != C.STATUS_CONFIRMED) {
+      if (_state != RtcSessionState.confirmed) {
         return false;
       }
 
@@ -2001,7 +2018,7 @@ class RTCSession extends EventManager implements Owner {
 
     try {
       RTCSessionDescription desc = await _processInDialogSdpOffer(request);
-      if (_status == C.STATUS_TERMINATED) return;
+      if (_state == RtcSessionState.terminated) return;
       // Send answer.
       sendAnswer(desc.sdp);
     } catch (error) {
@@ -2074,7 +2091,7 @@ class RTCSession extends EventManager implements Owner {
           }
         }
         emit(EventStream(
-            session: this, originator: 'local', stream: localStream));
+            session: this, originator: Originator.local, stream: localStream));
       } else {
         logger.w(
             'Remote wants to upgrade to video but no camera available to send');
@@ -2083,11 +2100,13 @@ class RTCSession extends EventManager implements Owner {
 
     logger.d('emit "sdp"');
     final String? processedSDP = _sdpOfferToWebRTC(request.body);
-    emit(EventSdp(originator: 'remote', type: 'offer', sdp: processedSDP));
+    emit(EventSdp(
+        originator: Originator.remote, type: SdpType.offer, sdp: processedSDP));
 
-    RTCSessionDescription offer = RTCSessionDescription(processedSDP, 'offer');
+    RTCSessionDescription offer =
+        RTCSessionDescription(processedSDP, SdpType.offer.name);
 
-    if (_status == C.STATUS_TERMINATED) {
+    if (_state == RtcSessionState.terminated) {
       throw Exceptions.InvalidStateError('terminated');
     }
     try {
@@ -2103,26 +2122,27 @@ class RTCSession extends EventManager implements Owner {
           'peerconnection.setRemoteDescription() failed');
     }
 
-    if (_status == C.STATUS_TERMINATED) {
+    if (_state == RtcSessionState.terminated) {
       throw Exceptions.InvalidStateError('terminated');
     }
 
     if (_remoteHold == true && hold == false) {
       _remoteHold = false;
-      _onunhold('remote');
+      _onunhold(Originator.remote);
     } else if (_remoteHold == false && hold == true) {
       _remoteHold = true;
-      _onhold('remote');
+      _onhold(Originator.remote);
     }
 
     // Create local description.
 
-    if (_status == C.STATUS_TERMINATED) {
+    if (_state == RtcSessionState.terminated) {
       throw Exceptions.InvalidStateError('terminated');
     }
 
     try {
-      return await _createLocalDescription('answer', _rtcAnswerConstraints);
+      return await _createLocalDescription(
+          SdpType.answer, _rtcAnswerConstraints);
     } catch (_) {
       request.reply(500);
       throw Exceptions.TypeError('_createLocalDescription() failed');
@@ -2157,8 +2177,8 @@ class RTCSession extends EventManager implements Owner {
         InitSuccessCallback? initCallback, Map<String, dynamic> options) {
       initCallback = (initCallback is Function) ? initCallback : null;
 
-      if (_status != C.STATUS_WAITING_FOR_ACK &&
-          _status != C.STATUS_CONFIRMED) {
+      if (_state != RtcSessionState.waitingForAck &&
+          _state != RtcSessionState.confirmed) {
         return false;
       }
 
@@ -2266,8 +2286,8 @@ class RTCSession extends EventManager implements Owner {
     logger.d('receiveReplaces()');
 
     bool accept(InitSuccessCallback initCallback) {
-      if (_status != C.STATUS_WAITING_FOR_ACK &&
-          _status != C.STATUS_CONFIRMED) {
+      if (_state != RtcSessionState.waitingForAck &&
+          _state != RtcSessionState.confirmed) {
         return false;
       }
 
@@ -2334,20 +2354,22 @@ class RTCSession extends EventManager implements Owner {
     // A stream is given, var the app set events such as 'peerconnection' and 'connecting'.
     if (mediaStream != null) {
       stream = mediaStream;
-      emit(EventStream(session: this, originator: 'local', stream: stream));
+      emit(EventStream(
+          session: this, originator: Originator.local, stream: stream));
     } // Request for user media access.
     else if (mediaConstraints['audio'] != null ||
         mediaConstraints['video'] != null) {
       _localMediaStreamLocallyGenerated = true;
       try {
         stream = await navigator.mediaDevices.getUserMedia(mediaConstraints);
-        emit(EventStream(session: this, originator: 'local', stream: stream));
+        emit(EventStream(
+            session: this, originator: Originator.local, stream: stream));
       } catch (error) {
-        if (_status == C.STATUS_TERMINATED) {
+        if (_state == RtcSessionState.terminated) {
           throw Exceptions.InvalidStateError('terminated');
         }
         _failed(
-            'local',
+            Originator.local,
             null,
             null,
             null,
@@ -2356,11 +2378,11 @@ class RTCSession extends EventManager implements Owner {
             'User Denied Media Access');
         logger.e('emit "getusermediafailed" [error:${error.toString()}]');
         emit(EventGetUserMediaFailed(exception: error));
-        throw error;
+        rethrow;
       }
     }
 
-    if (_status == C.STATUS_TERMINATED) {
+    if (_state == RtcSessionState.terminated) {
       throw Exceptions.InvalidStateError('terminated');
     }
 
@@ -2386,13 +2408,13 @@ class RTCSession extends EventManager implements Owner {
     _connecting(_request);
     try {
       RTCSessionDescription desc =
-          await _createLocalDescription('offer', rtcOfferConstraints);
-      if (_is_canceled || _status == C.STATUS_TERMINATED) {
+          await _createLocalDescription(SdpType.offer, rtcOfferConstraints);
+      if (_is_canceled || _state == RtcSessionState.terminated) {
         throw Exceptions.InvalidStateError('terminated');
       }
 
       _request.body = desc.sdp;
-      _status = C.STATUS_INVITE_SENT;
+      _state = RtcSessionState.inviteSent;
 
       logger.d('emit "sending" [request]');
 
@@ -2402,19 +2424,19 @@ class RTCSession extends EventManager implements Owner {
       request_sender.send();
     } catch (error, s) {
       logger.e(error.toString(), error: error, stackTrace: s);
-      _failed('local', null, null, null, 500, DartSIP_C.CausesType.WEBRTC_ERROR,
-          'Can\'t create local SDP');
-      if (_status == C.STATUS_TERMINATED) {
+      _failed(Originator.local, null, null, null, 500,
+          DartSIP_C.CausesType.WEBRTC_ERROR, 'Can\'t create local SDP');
+      if (_state == RtcSessionState.terminated) {
         return;
       }
       logger.e('Failed to _sendInitialRequest: ${error.toString()}');
-      throw error;
+      rethrow;
     }
   }
 
   /// Reception of Response for Initial INVITE
   void _receiveInviteResponse(IncomingResponse? response) async {
-    logger.d('receiveInviteResponse()  current status: $_status ');
+    logger.d('receiveInviteResponse()  current status: $_state ');
 
     if (response == null) {
       logger.d('No response received');
@@ -2459,7 +2481,8 @@ class RTCSession extends EventManager implements Owner {
       return;
     }
 
-    if (_status != C.STATUS_INVITE_SENT && _status != C.STATUS_1XX_RECEIVED) {
+    if (_state != RtcSessionState.inviteSent &&
+        _state != RtcSessionState.provisionalResponse) {
       return;
     }
 
@@ -2467,7 +2490,7 @@ class RTCSession extends EventManager implements Owner {
 
     if (utils.test100(status_code)) {
       // 100 trying
-      _status = C.STATUS_1XX_RECEIVED;
+      _state = RtcSessionState.provisionalResponse;
     } else if (utils.test1XX(status_code)) {
       // 1XX
       // Do nothing with 1xx responses without To tag.
@@ -2484,18 +2507,21 @@ class RTCSession extends EventManager implements Owner {
         }
       }
 
-      _status = C.STATUS_1XX_RECEIVED;
-      _progress('remote', response, int.parse(status_code));
+      _state = RtcSessionState.provisionalResponse;
+      _progress(Originator.remote, response, int.parse(status_code));
 
       if (response.body == null || response.body!.isEmpty) {
         return;
       }
 
       logger.d('emit "sdp"');
-      emit(EventSdp(originator: 'remote', type: 'answer', sdp: response.body));
+      emit(EventSdp(
+          originator: Originator.remote,
+          type: SdpType.answer,
+          sdp: response.body));
 
       RTCSessionDescription answer =
-          RTCSessionDescription(response.body, 'answer');
+          RTCSessionDescription(response.body, SdpType.answer.name);
 
       try {
         await _connection!.setRemoteDescription(answer);
@@ -2506,11 +2532,11 @@ class RTCSession extends EventManager implements Owner {
       }
     } else if (utils.test2XX(status_code)) {
       // 2XX
-      _status = C.STATUS_CONFIRMED;
+      _state = RtcSessionState.confirmed;
 
       if (response.body == null || response.body!.isEmpty) {
         _acceptAndTerminate(response, 400, DartSIP_C.CausesType.MISSING_SDP);
-        _failed('remote', null, null, response, 400,
+        _failed(Originator.remote, null, null, response, 400,
             DartSIP_C.CausesType.BAD_MEDIA_DESCRIPTION, 'Missing SDP');
         return;
       }
@@ -2519,7 +2545,7 @@ class RTCSession extends EventManager implements Owner {
 
       if (mediaPort == 0 && _ua.configuration.terminateOnAudioMediaPortZero) {
         _acceptAndTerminate(response, 400, DartSIP_C.CausesType.MISSING_SDP);
-        _failed('remote', null, null, response, 400,
+        _failed(Originator.remote, null, null, response, 400,
             DartSIP_C.CausesType.BAD_MEDIA_DESCRIPTION, 'Media port is zero');
         return;
       }
@@ -2530,10 +2556,13 @@ class RTCSession extends EventManager implements Owner {
       }
 
       logger.d('emit "sdp"');
-      emit(EventSdp(originator: 'remote', type: 'answer', sdp: response.body));
+      emit(EventSdp(
+          originator: Originator.remote,
+          type: SdpType.answer,
+          sdp: response.body));
 
       RTCSessionDescription answer =
-          RTCSessionDescription(response.body, 'answer');
+          RTCSessionDescription(response.body, SdpType.answer.name);
 
       // Be ready for 200 with SDP after a 180/183 with SDP.
       // We created a SDP 'answer' for it, so check the current signaling state.
@@ -2548,7 +2577,7 @@ class RTCSession extends EventManager implements Owner {
         } catch (error) {
           _acceptAndTerminate(response, 500, error.toString());
           _failed(
-              'local',
+              Originator.local,
               null,
               null,
               response,
@@ -2562,12 +2591,12 @@ class RTCSession extends EventManager implements Owner {
         await _connection!.setRemoteDescription(answer);
         // Handle Session Timers.
         _handleSessionTimersInIncomingResponse(response);
-        _accepted('remote', response);
+        _accepted(Originator.remote, response);
         OutgoingRequest ack = sendRequest(SipMethod.ACK);
-        _confirmed('local', ack);
+        _confirmed(Originator.local, ack);
       } catch (error) {
         _acceptAndTerminate(response, 488, 'Not Acceptable Here');
-        _failed('remote', null, null, response, 488,
+        _failed(Originator.remote, null, null, response, 488,
             DartSIP_C.CausesType.BAD_MEDIA_DESCRIPTION, 'Not Acceptable Here');
         logger.e(
             'emit "peerconnection:setremotedescriptionfailed" [error:${error.toString()}]');
@@ -2575,8 +2604,8 @@ class RTCSession extends EventManager implements Owner {
       }
     } else {
       String cause = utils.sipErrorCause(response.status_code);
-      _failed('remote', null, null, response, response.status_code, cause,
-          response.reason_phrase);
+      _failed(Originator.remote, null, null, response, response.status_code,
+          cause, response.reason_phrase);
     }
   }
 
@@ -2611,7 +2640,7 @@ class RTCSession extends EventManager implements Owner {
     }
 
     void onSucceeded(IncomingResponse? response) async {
-      if (_status == C.STATUS_TERMINATED) {
+      if (_state == RtcSessionState.terminated) {
         return;
       }
 
@@ -2635,10 +2664,14 @@ class RTCSession extends EventManager implements Owner {
       }
 
       logger.d('emit "sdp"');
-      emit(EventSdp(originator: 'remote', type: 'answer', sdp: response.body));
+      emit(EventSdp(
+        originator: Originator.remote,
+        type: SdpType.answer,
+        sdp: response.body,
+      ));
 
       RTCSessionDescription answer =
-          RTCSessionDescription(response.body, 'answer');
+          RTCSessionDescription(response.body, SdpType.answer.name);
 
       try {
         await _connection!.setRemoteDescription(answer);
@@ -2653,10 +2686,11 @@ class RTCSession extends EventManager implements Owner {
 
     try {
       RTCSessionDescription desc =
-          await _createLocalDescription('offer', rtcOfferConstraints);
+          await _createLocalDescription(SdpType.offer, rtcOfferConstraints);
       String? sdp = _mangleOffer(desc.sdp);
       logger.d('emit "sdp"');
-      emit(EventSdp(originator: 'local', type: 'offer', sdp: sdp));
+      emit(EventSdp(
+          originator: Originator.local, type: SdpType.offer, sdp: sdp));
 
       EventManager handlers = EventManager();
       handlers.on(EventOnSuccessResponse(), (EventOnSuccessResponse event) {
@@ -2732,14 +2766,16 @@ class RTCSession extends EventManager implements Owner {
       }
 
       emit(EventStream(
-          session: this, originator: 'local', stream: _localMediaStream));
+          session: this,
+          originator: Originator.local,
+          stream: _localMediaStream));
     } catch (error) {
-      if (_status == C.STATUS_TERMINATED) {
+      if (_state == RtcSessionState.terminated) {
         throw Exceptions.InvalidStateError('terminated');
       }
       request.reply(480);
       _failed(
-          'local',
+          Originator.local,
           null,
           null,
           null,
@@ -2767,7 +2803,7 @@ class RTCSession extends EventManager implements Owner {
     }
 
     void onSucceeded(IncomingResponse? response) async {
-      if (_status == C.STATUS_TERMINATED) {
+      if (_state == RtcSessionState.terminated) {
         return;
       }
 
@@ -2791,10 +2827,13 @@ class RTCSession extends EventManager implements Owner {
       }
 
       logger.d('emit "sdp"');
-      emit(EventSdp(originator: 'remote', type: 'answer', sdp: response.body));
+      emit(EventSdp(
+          originator: Originator.remote,
+          type: SdpType.answer,
+          sdp: response.body));
 
       RTCSessionDescription answer =
-          RTCSessionDescription(response.body, 'answer');
+          RTCSessionDescription(response.body, SdpType.answer.name);
 
       try {
         await _connection!.setRemoteDescription(answer);
@@ -2809,10 +2848,11 @@ class RTCSession extends EventManager implements Owner {
 
     try {
       RTCSessionDescription desc =
-          await _createLocalDescription('offer', rtcOfferConstraints);
+          await _createLocalDescription(SdpType.offer, rtcOfferConstraints);
       String? sdp = _mangleOffer(desc.sdp);
       logger.d('emit "sdp"');
-      emit(EventSdp(originator: 'local', type: 'offer', sdp: sdp));
+      emit(EventSdp(
+          originator: Originator.local, type: SdpType.offer, sdp: sdp));
 
       EventManager handlers = EventManager();
       handlers.on(EventOnSuccessResponse(), (EventOnSuccessResponse event) {
@@ -2874,7 +2914,7 @@ class RTCSession extends EventManager implements Owner {
     }
 
     void onSucceeded(IncomingResponse? response) async {
-      if (_status == C.STATUS_TERMINATED) {
+      if (_state == RtcSessionState.terminated) {
         return;
       }
 
@@ -2897,11 +2937,13 @@ class RTCSession extends EventManager implements Owner {
         }
 
         logger.d('emit "sdp"');
-        emit(
-            EventSdp(originator: 'remote', type: 'answer', sdp: response.body));
+        emit(EventSdp(
+            originator: Originator.remote,
+            type: SdpType.answer,
+            sdp: response.body));
 
         RTCSessionDescription answer =
-            RTCSessionDescription(response.body, 'answer');
+            RTCSessionDescription(response.body, SdpType.answer.name);
 
         try {
           await _connection!.setRemoteDescription(answer);
@@ -2923,11 +2965,12 @@ class RTCSession extends EventManager implements Owner {
       extraHeaders.add('Content-Type: application/sdp');
       try {
         RTCSessionDescription desc =
-            await _createLocalDescription('offer', rtcOfferConstraints);
+            await _createLocalDescription(SdpType.offer, rtcOfferConstraints);
         String? sdp = _mangleOffer(desc.sdp);
 
         logger.d('emit "sdp"');
-        emit(EventSdp(originator: 'local', type: 'offer', sdp: sdp));
+        emit(EventSdp(
+            originator: Originator.local, type: SdpType.offer, sdp: sdp));
 
         EventManager handlers = EventManager();
         handlers.on(EventOnSuccessResponse(), (EventOnSuccessResponse event) {
@@ -3004,7 +3047,7 @@ class RTCSession extends EventManager implements Owner {
     }
 
     // Update session status.
-    _status = C.STATUS_TERMINATED;
+    _state = RtcSessionState.terminated;
   }
 
   /**
@@ -3170,7 +3213,7 @@ class RTCSession extends EventManager implements Owner {
     // I'm the refresher.
     if (_sessionTimers.refresher) {
       _sessionTimers.timer = setTimeout(() {
-        if (_status == C.STATUS_TERMINATED) {
+        if (_state == RtcSessionState.terminated) {
           return;
         }
 
@@ -3186,7 +3229,7 @@ class RTCSession extends EventManager implements Owner {
     // I'm not the refresher.
     else {
       _sessionTimers.timer = setTimeout(() {
-        if (_status == C.STATUS_TERMINATED) {
+        if (_state == RtcSessionState.terminated) {
           return;
         }
 
@@ -3217,7 +3260,7 @@ class RTCSession extends EventManager implements Owner {
     }
   }
 
-  void _newRTCSession(String originator, dynamic request) {
+  void _newRTCSession(Originator originator, dynamic request) {
     logger.d('newRTCSession()');
     _ua.newRTCSession(originator: originator, session: this, request: request);
   }
@@ -3228,7 +3271,7 @@ class RTCSession extends EventManager implements Owner {
     emit(EventCallConnecting(session: this, request: request));
   }
 
-  void _progress(String originator, dynamic response, [int? status_code]) {
+  void _progress(Originator originator, dynamic response, [int? status_code]) {
     logger.d('session progress');
     logger.d('emit "progress"');
 
@@ -3241,7 +3284,7 @@ class RTCSession extends EventManager implements Owner {
         cause: errorCause));
   }
 
-  void _accepted(String originator, [dynamic message]) {
+  void _accepted(Originator originator, [dynamic message]) {
     logger.d('session accepted');
     _start_time = DateTime.now();
     logger.d('emit "accepted"');
@@ -3249,14 +3292,15 @@ class RTCSession extends EventManager implements Owner {
         session: this, originator: originator, response: message));
   }
 
-  void _confirmed(String originator, dynamic ack) {
+  void _confirmed(Originator originator, dynamic ack) {
     logger.d('session confirmed');
     _is_confirmed = true;
     logger.d('emit "confirmed"');
     emit(EventCallConfirmed(session: this, originator: originator, ack: ack));
   }
 
-  void _ended(String originator, IncomingRequest? request, ErrorCause cause) {
+  void _ended(
+      Originator originator, IncomingRequest? request, ErrorCause cause) {
     logger.d('session ended');
     _end_time = DateTime.now();
     _close();
@@ -3265,7 +3309,7 @@ class RTCSession extends EventManager implements Owner {
         session: this, originator: originator, request: request, cause: cause));
   }
 
-  void _failed(String originator, dynamic message, dynamic request,
+  void _failed(Originator originator, dynamic message, dynamic request,
       dynamic response, int? status_code, String cause, String? reason_phrase) {
     logger.d('session failed');
 
@@ -3290,14 +3334,14 @@ class RTCSession extends EventManager implements Owner {
         response: response));
   }
 
-  void _onhold(String originator) {
+  void _onhold(Originator originator) {
     logger.d('session onhold');
     _setLocalMediaStatus();
     logger.d('emit "hold"');
     emit(EventCallHold(session: this, originator: originator));
   }
 
-  void _onunhold(String originator) {
+  void _onunhold(Originator originator) {
     logger.d('session onunhold');
     _setLocalMediaStatus();
     logger.d('emit "unhold"');
