@@ -1,16 +1,34 @@
 import 'dart:async';
 import 'dart:math';
 
-import 'package:sip_ua/src/event_manager/events.dart';
-import 'package:sip_ua/src/transport_constants.dart';
-import 'package:sip_ua/src/transports/socket_interface.dart';
-import 'package:sip_ua/src/transports/tcp_socket.dart';
+import './event_manager/events.dart';
+import './transports/socket_interface.dart';
 import 'exceptions.dart' as Exceptions;
 import 'logger.dart';
 import 'stack_trace_nj.dart';
 import 'timers.dart';
-import 'transports/web_socket.dart';
 import 'utils.dart';
+
+enum TransportStatus { connected, connecting, disconnected }
+
+enum SocketStatus { ready, error }
+
+class SocketInfo {
+  SocketInfo({
+    required this.socket,
+    required this.weight,
+    required this.status,
+  });
+
+  SIPUASocketInterface socket;
+  int weight;
+  SocketStatus status;
+}
+
+const Map<String, int> defaultRecoveryOptions = <String, int>{
+  'min_interval': 2, // minimum interval in seconds between recover attempts
+  'max_interval': 30 // maximum interval in seconds between recover attempts
+};
 
 /*
  * Manages one or multiple DartSIP.Socket instances.
@@ -19,34 +37,36 @@ import 'utils.dart';
  * @socket DartSIP::Socket instance
  */
 class SocketTransport {
-  SocketTransport(List<SIPUASocketInterface>? sockets,
-      [Map<String, int> recovery_options = C.recovery_options]) {
+  SocketTransport(
+    List<SIPUASocketInterface>? sockets, [
+    Map<String, int> recovery_options = defaultRecoveryOptions,
+  ]) {
     logger.d('Socket Transport new()');
 
     _recovery_options = recovery_options;
 
     // We must recieve at least 1 socket
-    if (sockets!.length == 0) {
+    if (sockets!.isEmpty) {
       throw Exceptions.TypeError(
           'invalid argument: Must recieve atleast 1 web socket');
     }
 
-    for (SIPUASocketInterface socket in sockets) {
-      _socketsMap.add(<String, dynamic>{
-        'socket': socket,
-        'weight': socket.weight ?? 0,
-        'status': C.SOCKET_STATUS_READY
-      });
+    for (final SIPUASocketInterface socket in sockets) {
+      _sockets.add(SocketInfo(
+        socket: socket,
+        weight: socket.weight ?? 0,
+        status: SocketStatus.ready,
+      ));
     }
     // Get the socket with higher weight.
     _getSocket();
   }
 
-  int status = C.STATUS_DISCONNECTED;
+  TransportStatus status = TransportStatus.disconnected;
   // Current socket.
   late SIPUASocketInterface socket;
   // Socket collection.
-  final List<Map<String, dynamic>> _socketsMap = <Map<String, dynamic>>[];
+  final List<SocketInfo> _sockets = <SocketInfo>[];
   late Map<String, int> _recovery_options;
   int _recover_attempts = 0;
   Timer? _recovery_timer;
@@ -82,7 +102,7 @@ class SocketTransport {
     }
 
     _close_requested = false;
-    status = C.STATUS_CONNECTING;
+    status = TransportStatus.connecting;
     onconnecting(socket, _recover_attempts);
 
     if (!_close_requested) {
@@ -100,7 +120,7 @@ class SocketTransport {
 
     _close_requested = true;
     _recover_attempts = 0;
-    status = C.STATUS_DISCONNECTED;
+    status = TransportStatus.disconnected;
 
     // Clear recovery_timer.
     if (_recovery_timer != null) {
@@ -117,11 +137,10 @@ class SocketTransport {
 
     socket.disconnect();
     ondisconnect(
-        socket,
-        ErrorCause(
-            cause: 'disconnect',
-            status_code: 0,
-            reason_phrase: 'close by local'));
+      socket,
+      ErrorCause(
+          cause: 'disconnect', status_code: 0, reason_phrase: 'close by local'),
+    );
   }
 
   bool send(dynamic data) {
@@ -139,11 +158,11 @@ class SocketTransport {
   }
 
   bool isConnected() {
-    return status == C.STATUS_CONNECTED;
+    return status == TransportStatus.connected;
   }
 
   bool isConnecting() {
-    return status == C.STATUS_CONNECTING;
+    return status == TransportStatus.connecting;
   }
 
   /**
@@ -180,28 +199,28 @@ class SocketTransport {
   void _getSocket() {
     // If we dont have at least 1 socket to try and use, thiw will loop endlessly
 
-    if (_socketsMap.length == 0) {
+    if (_sockets.isEmpty) {
       throw Exceptions.TypeError('invalid argument: too few sockets');
     }
 
-    List<Map<String, dynamic>> candidates = <Map<String, dynamic>>[];
+    List<SocketInfo> candidates = <SocketInfo>[];
 
-    for (Map<String, dynamic> socket in _socketsMap) {
-      if (socket['status'] == C.SOCKET_STATUS_ERROR) {
+    for (final SocketInfo socket in _sockets) {
+      if (socket.status == SocketStatus.error) {
         return; // continue the array iteration
       } else if (candidates.isEmpty) {
         candidates.add(socket);
-      } else if (socket['weight'] > candidates[0]['weight']) {
-        candidates = <Map<String, dynamic>>[socket];
-      } else if (socket['weight'] == candidates[0]['weight']) {
+      } else if (socket.weight > candidates[0].weight) {
+        candidates = <SocketInfo>[socket];
+      } else if (socket.weight == candidates[0].weight) {
         candidates.add(socket);
       }
     }
 
     if (candidates.isEmpty) {
       // All sockets have failed. reset sockets status.
-      for (Map<String, dynamic> socket in _socketsMap) {
-        socket['status'] = C.SOCKET_STATUS_READY;
+      for (final SocketInfo socket in _sockets) {
+        socket.status = SocketStatus.ready;
       }
       // Get next available socket.
       _getSocket();
@@ -210,7 +229,7 @@ class SocketTransport {
 
     num idx = (Math.randomDouble() * candidates.length).floor();
 
-    socket = candidates[idx as int]['socket'];
+    socket = candidates[idx as int].socket;
   }
 
   /**
@@ -219,7 +238,7 @@ class SocketTransport {
 
   void _onConnect() {
     _recover_attempts = 0;
-    status = C.STATUS_CONNECTED;
+    status = TransportStatus.connected;
 
     // Clear recovery_timer.
     if (_recovery_timer != null) {
@@ -231,7 +250,7 @@ class SocketTransport {
 
   void _onDisconnect(
       SIPUASocketInterface socket, bool error, int? closeCode, String? reason) {
-    status = C.STATUS_DISCONNECTED;
+    status = TransportStatus.disconnected;
     ondisconnect(
         socket,
         ErrorCause(
@@ -242,9 +261,9 @@ class SocketTransport {
     }
     // Update socket status.
     else {
-      for (Map<String, dynamic> socket in _socketsMap) {
-        if (socket == socket['socket']) {
-          socket['status'] = C.SOCKET_STATUS_ERROR;
+      for (final SocketInfo s in _sockets) {
+        if (socket == s.socket) {
+          s.status = SocketStatus.error;
         }
       }
     }
@@ -266,8 +285,9 @@ class SocketTransport {
         data = String.fromCharCodes(data);
       } catch (evt) {
         logger.d(
-            'received binary message [${data.runtimeType}]failed to be converted into string,'
-            ' message discarded');
+          'received binary message [${data.runtimeType}]failed to be converted into string,'
+          ' message discarded',
+        );
         return;
       }
       logger.d('received binary message:\n\n$data\n');
