@@ -1,5 +1,5 @@
-import 'package:sip_ua/src/sip_message.dart';
 import '../constants.dart';
+import '../sip_message.dart';
 import '../timers.dart';
 import 'invite_server.dart';
 import 'non_invite_server.dart';
